% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_country.R
\name{compare_simdata}
\alias{compare_simdata}
\title{Compare simulated trajectory with estimated parameters to empirical data}
\usage{
compare_simdata(country_data, contact_data, age_data, Np = 100, Niter = 1e+05)
}
\arguments{
\item{country_data}{Epidemiological dataset}

\item{contact_data}{Contact dataset}

\item{age_data}{Age dataset}

\item{Np}{Integer. Number of simulations. This also corresponds to the number
of particles in the particle filter.}

\item{Niter}{Integer. Number of iterations in the MCMC when \verb{task = "estimate}}
}
\description{
Compare simulated trajectory with estimated parameters to empirical data
}
