% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_asympto.R
\name{estimate_asympto}
\alias{estimate_asympto}
\title{Estimate the proportion of asymptomatic cases by capture/recapture}
\usage{
estimate_asympto(country_data, smooth = 1)
}
\arguments{
\item{country_data}{The output of \code{\link[=load_country_data]{load_country_data()}}}

\item{smooth}{Odd integer. Number of days on which smoothing for PropAsympto
should be performed. Defaults to 1 (no smoothing).}
}
\description{
Estimate the proportion of asymptomatic cases by capture/recapture
}
