% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHS_init.R
\name{LHS_init}
\alias{LHS_init}
\title{Run particle filter or stochastic simulations}
\usage{
LHS_init(
  country_data,
  contact_data,
  age_data,
  Np = 20,
  Niter = 1e+06,
  task = c("estimate", "run"),
  transmRate,
  vecEff,
  R0
)
}
\arguments{
\item{country_data}{Epidemiological dataset}

\item{contact_data}{Contact dataset}

\item{age_data}{Age dataset}

\item{Np}{Integer. Number of simulations. This also corresponds to the number
of particles in the particle filter.}

\item{Niter}{Integer. Number of iterations in the MCMC when \verb{task = "estimate}}

\item{task}{The task to perform. Either \code{"estimate"} to estimate parameters
or \verb{"run} to run simulations (\code{transmRate}, and \code{vecEff} must
be provided in this case).}

\item{transmRate, vecEff, R0}{Epidemiological parameters. Only one of
\code{transmRate} and \code{R0} can be provided.}

\item{prop_asympto}{Proportion of undetected cases}
}
\description{
Run particle filter or stochastic simulations
}
