% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asympto.R
\name{plot_asympto}
\alias{plot_asympto}
\title{Plot the proportion of asymptomatic}
\usage{
plot_asympto(country_data)
}
\arguments{
\item{country_data}{the result of \code{\link[=estimate_asympto]{estimate_asympto()}}}
}
\description{
Plot the proportion of asymptomatic
}
