% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulations.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Plot both the simulations from estimated parameters and empirical data}
\usage{
plot_simulations(outreprod, country_data)
}
\arguments{
\item{outreprod}{Result of the simulations}

\item{country_data}{Epidemiological dataset}
}
\description{
Plot both the simulations from estimated parameters and empirical data
}
