% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_estimation.R
\name{summarise_estimation}
\alias{summarise_estimation}
\title{Get 95\% credibility intervals for estimations}
\usage{
summarise_estimation(MCMC, country_data)
}
\arguments{
\item{MCMC}{the result of \code{run_MCMC()}}

\item{country_data}{Epidemiological dataset}
}
\description{
Get 95\% credibility intervals for estimations
}
